Add provisioning for all VM groups and volumes
Create etcd cluster, master VMs, load balancer VMs, NFS VMs and SSD
nodes. Add hosts in each group of hosts to a server group with an
anti-affinity scheduling rule. Write a type for creating a VM with a
volume attached and use that type in the resource groups for LB and NFS
hosts.