Provisioning for a bastion host
Add a bastion host to the Heat stack with a security group that is only
assigned to the bastion host and that allows SSH access from this host.
External SSH access is blocked for all other hosts - access to those is
via the bastion.

Assign a public IP address to the bastion similarly as the OpenShift IP
address is assigned, i.e. with Ansible instead of Heat. This allows us
to keep a static IP for both the bastion host and the OpenShift service.