Support inserting cloud-init config for bastion
In order to make the bastion host ready to act as a bastion host, add
the possibility to add some initial configuration via cloud-init. This
can be used for e.g. installing netcat that is required for proxying to
hosts with no public IP. It can also later be used to insert SSH host
keys.

Also set StrictHostKeyChecking to no for the bastion for now to avoid
having to run ssh-keygen -R every time it is reprovisioned. This could
be changed if a host key for the bastion is deployed instead.