Always create dedicated net+router for OpenShift
Modify the Heat template for the OpenShift cluster so that a dedicated
router and network is always created. Use the defaults when creating
these dedicated resources via os_stack.

Use 192.168.0.0/16 as the default CIDR for the dedicated network. This
is hardcoded in roles/openshift_node/tasks/main.yml for firewall rules,
so we can't use e.g. 10.10.0.0/16 without changing the openshift_node
role first. Changing the default for the dedicated cluster network CIDR
is the easier solution for now. The hardcoded firewall rules in the
openshift_node role can be changed if a need arises.

To allow clean deletion of the stack, add dependencies from the VMs and
the router to the subnet that is created. This make stack deletions
work presumably because the dependent resources are deleted first and
only then is the subnet deleted.

Add a playbook for deprovisioning the Heat stack. It is necessary to
disassociate floating IPs from VMs first before deleting the Heat stack
for OpenShift when floating IPs are managed outside the Heat stack and
the Heat stack creates a new cluster network and subnet. Write a
playbook to do this so that the user does not have to manually
disassociate the floating IPs first.