containerized provisioning environment
This commit adds pac-deployer, a container with all external software
dependencies included for running deployment. Scripts for local builds
as well as creating a build config in existing OpenShift are included
(the latter is a consolidated version of existing build config script
for oso-default-www). There is also a script to launch a deployment
container.

The documentation in README.md is updated to include running
containerized deployment for heat_provisioning playbook.

Master copy of python dependency file requirements.txt is now held in
container-src/pac-deployer/ and a symlink exists for backwards
compatibility in playbooks/openshift/.