support deployment of default www application
Added deployment code and sources for an application serving a page
with links to the cluster management api and developer documentation.
This is intended to serve default locations for the cluster, like
oso.example.org and www.oso.example.org for a less confusing user
experience when trying to find the admin console or basic information.

Container source code is located under /src/oso-default-www/ along
with a script to build it.

Deployment is done in post_install -playbook and uses the following
variables:

  * deploy_default_www_app - set to true to enable deployment
  * default_www_app_image  - set to docker registry uri for the
                             application image